requirements.txt
streamlit>=1.44,<2
pandas>=2.2,<3
openpyxl>=3.1,<4
python-dateutil>=2.8
